refactor: renamed entity file
