+++ src/domain/Entity.ts
