+++ tests/Entity.spec.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
-import Entity from '../src/domain/entity';
+import Entity from '../src/domain/Entity';
 import EntityDiscardedError from '../src/domain/exceptions/EntityDiscardedError';
 
 class MyEntity extends Entity<string> {
